+++ src/main/java/web/lunarsonic/models/Result.java
@@ -1,4 +1,4 @@
 package web.lunarsonic.models;
 
-public record Result(float x, int y, float r, boolean hit, long scriptTime, String serverTime) {
+public record Result(float x, float y, float r, boolean hit, long scriptTime, String serverTime) {
 }
